# docs/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LecheFacil Backend",
    "description": "Multi-tenant API for LecheFacil MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dashboard/daily-kpis": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Daily Kpis",
        "operationId": "get_daily_kpis_api_v1_dashboard_daily_kpis_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyKPIsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/top-producers": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Top Producers",
        "operationId": "get_top_producers_api_v1_dashboard_top_producers_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopProducersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/daily-progress": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Daily Progress",
        "operationId": "get_daily_progress_api_v1_dashboard_daily_progress_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/alerts": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Alerts",
        "operationId": "get_alerts_api_v1_dashboard_alerts_get",
        "parameters": [
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "high",
                "medium",
                "low"
              ],
              "type": "string",
              "default": "all",
              "title": "Priority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/worker-progress": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Worker Progress",
        "operationId": "get_worker_progress_api_v1_dashboard_worker_progress_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/vet-alerts": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Vet Alerts",
        "operationId": "get_vet_alerts_api_v1_dashboard_vet_alerts_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetAlertsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/admin-overview": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Admin Overview",
        "operationId": "get_admin_overview_api_v1_dashboard_admin_overview_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminManagementOverview": {
        "properties": {
          "monthly_profitability": {
            "type": "string",
            "title": "Monthly Profitability"
          },
          "production_vs_goal": {
            "type": "string",
            "title": "Production Vs Goal"
          },
          "pending_alerts": {
            "type": "integer",
            "title": "Pending Alerts"
          },
          "upcoming_tasks": {
            "type": "integer",
            "title": "Upcoming Tasks"
          }
        },
        "type": "object",
        "required": [
          "monthly_profitability",
          "production_vs_goal",
          "pending_alerts",
          "upcoming_tasks"
        ],
        "title": "AdminManagementOverview"
      },
      "AdminOverviewResponse": {
        "properties": {
          "management_overview": {
            "$ref": "#/components/schemas/AdminManagementOverview"
          }
        },
        "type": "object",
        "required": [
          "management_overview"
        ],
        "title": "AdminOverviewResponse"
      },
      "AlertsResponse": {
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/DashboardAlert"
            },
            "type": "array",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "alerts"
        ],
        "title": "AlertsResponse"
      },
...
      "DailyGoal": {
        "properties": {
          "target_liters": {
            "type": "string",
            "title": "Target Liters"
          },
          "current_liters": {
            "type": "string",
            "title": "Current Liters"
          },
          "completion_percentage": {
            "type": "string",
            "title": "Completion Percentage"
          }
        },
        "type": "object",
        "required": [
          "target_liters",
          "current_liters",
          "completion_percentage"
        ],
        "title": "DailyGoal"
      },
      "DailyKPIsResponse": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "total_liters": {
            "type": "string",
            "title": "Total Liters"
          },
          "total_revenue": {
            "type": "string",
            "title": "Total Revenue"
          },
          "average_per_animal": {
            "type": "string",
            "title": "Average Per Animal"
          },
          "active_animals_count": {
            "type": "integer",
            "title": "Active Animals Count"
          },
          "trends": {
            "$ref": "#/components/schemas/DailyKPIsTrends"
          }
        },
        "type": "object",
        "required": [
          "date",
          "total_liters",
          "total_revenue",
          "average_per_animal",
          "active_animals_count",
          "trends"
        ],
        "title": "DailyKPIsResponse"
      },
      "DailyKPIsTrends": {
        "properties": {
          "liters_vs_yesterday": {
            "type": "string",
            "title": "Liters Vs Yesterday"
          },
          "revenue_vs_yesterday": {
            "type": "string",
            "title": "Revenue Vs Yesterday"
          },
          "average_vs_yesterday": {
            "type": "string",
            "title": "Average Vs Yesterday"
          }
        },
        "type": "object",
        "required": [
          "liters_vs_yesterday",
          "revenue_vs_yesterday",
          "average_vs_yesterday"
        ],
        "title": "DailyKPIsTrends"
      },
      "DailyProgressResponse": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "shifts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ShiftProgress"
            },
            "type": "object",
            "title": "Shifts"
          },
          "daily_goal": {
            "$ref": "#/components/schemas/DailyGoal"
          }
        },
        "type": "object",
        "required": [
          "date",
          "shifts",
          "daily_goal"
        ],
        "title": "DailyProgressResponse"
      },
      "DashboardAlert": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "health",
              "production",
              "price"
            ],
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Priority"
          },
          "animal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Animal Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "message",
          "priority",
          "created_at"
        ],
        "title": "DashboardAlert"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShiftProgress": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "in_progress",
              "pending"
            ],
            "title": "Status"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "liters": {
            "type": "string",
            "title": "Liters"
          }
        },
        "type": "object",
        "required": [
          "status",
          "liters"
        ],
        "title": "ShiftProgress"
      },
...
      "TopProducer": {
        "properties": {
          "animal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Animal Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "today_liters": {
            "type": "string",
            "title": "Today Liters"
          },
          "trend": {
            "type": "string",
            "enum": [
              "up",
              "down",
              "stable"
            ],
            "title": "Trend"
          },
          "trend_percentage": {
            "type": "string",
            "title": "Trend Percentage"
          }
        },
        "type": "object",
        "required": [
          "animal_id",
          "name",
          "tag",
          "today_liters",
          "trend",
          "trend_percentage"
        ],
        "title": "TopProducer"
      },
      "TopProducersResponse": {
        "properties": {
          "top_producers": {
            "items": {
              "$ref": "#/components/schemas/TopProducer"
            },
            "type": "array",
            "title": "Top Producers"
          }
        },
        "type": "object",
        "required": [
          "top_producers"
        ],
        "title": "TopProducersResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VetAlertsResponse": {
        "properties": {
          "health_summary": {
            "$ref": "#/components/schemas/VetHealthSummary"
          },
          "urgent_alerts": {
            "items": {
              "$ref": "#/components/schemas/VetUrgentAlert"
            },
            "type": "array",
            "title": "Urgent Alerts"
          }
        },
        "type": "object",
        "required": [
          "health_summary",
          "urgent_alerts"
        ],
        "title": "VetAlertsResponse"
      },
      "VetHealthSummary": {
        "properties": {
          "animals_in_treatment": {
            "type": "integer",
            "title": "Animals In Treatment"
          },
          "active_milk_withdrawals": {
            "type": "integer",
            "title": "Active Milk Withdrawals"
          },
          "upcoming_vaccinations": {
            "type": "integer",
            "title": "Upcoming Vaccinations"
          }
        },
        "type": "object",
        "required": [
          "animals_in_treatment",
          "active_milk_withdrawals",
          "upcoming_vaccinations"
        ],
        "title": "VetHealthSummary"
      },
      "VetUrgentAlert": {
        "properties": {
          "animal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Animal Id"
          },
          "animal_name": {
            "type": "string",
            "title": "Animal Name"
          },
          "animal_tag": {
            "type": "string",
            "title": "Animal Tag"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "type": "string",
            "title": "Details"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "animal_id",
          "animal_name",
          "animal_tag",
          "alert_type",
          "message",
          "details",
          "priority"
        ],
        "title": "VetUrgentAlert"
      },
      "WorkerProgress": {
        "properties": {
          "animals_milked": {
            "type": "integer",
            "title": "Animals Milked"
          },
          "total_animals_assigned": {
            "type": "integer",
            "title": "Total Animals Assigned"
          },
          "liters_recorded": {
            "type": "string",
            "title": "Liters Recorded"
          },
          "current_shift": {
            "type": "string",
            "enum": [
              "AM",
              "PM"
            ],
            "title": "Current Shift"
          },
          "shift_start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Shift Start Time"
          }
        },
        "type": "object",
        "required": [
          "animals_milked",
          "total_animals_assigned",
          "liters_recorded",
          "current_shift",
          "shift_start_time"
        ],
        "title": "WorkerProgress"
      },
      "WorkerProgressResponse": {
        "properties": {
          "today_progress": {
            "$ref": "#/components/schemas/WorkerProgress"
          }
        },
        "type": "object",
        "required": [
          "today_progress"
        ],
        "title": "WorkerProgressResponse"
      }
    }
  }
}